Move tests to appropriate package
